+++ test/ids-about/ids-about-percy-test.ts
@@ -1,4 +1,4 @@
-import percySnapshot from '@percy/puppeteer';
+import pageSnapshot from '../helpers/page-snapshot';
 
 describe('Ids About Percy Tests', () => {
   const url = 'http://localhost:4444/ids-about/example.html';
@@ -9,37 +9,37 @@ describe('Ids About Percy Tests', () => {
       (document.querySelector('#about-example-trigger') as any).click();
     });
     await page.waitForTimeout(200); // approx. time for a Modal to show
-    await percySnapshot(page, 'ids-about-new-light');
+    await pageSnapshot(page, 'ids-about-new-light');
   });
 
   it('should not have visual regressions in standalone css', async () => {
     await page.goto('http://localhost:4444/ids-about/standalone-css.html', { waitUntil: ['networkidle2', 'load'] });
-    await percySnapshot(page, 'ids-about-standalone-css', { widths: [1280] });
+    await pageSnapshot(page, 'ids-about-standalone-css', { widths: [1280] });
   });
 
-  it('should not have visual regressions in new dark theme (percy)', async () => {
+  it.skip('should not have visual regressions in new dark theme (percy)', async () => {
     await page.goto(url, { waitUntil: ['networkidle2', 'load'] });
     await page.evaluate(() => {
-      document.querySelector('ids-theme-switcher')?.setAttribute('mode', 'dark');
+      document.querySelector('ids-theme-switcher')?.setAttribute('theme', 'default-dark');
     });
     await page.waitForTimeout(200);
     await page.evaluate(() => {
       (document.querySelector('#about-example-trigger') as any).click();
     });
     await page.waitForTimeout(200); // approx. time for a Modal to show
-    await percySnapshot(page, 'ids-about-new-dark');
+    await pageSnapshot(page, 'ids-about-new-dark');
   });
 
-  it('should not have visual regressions in new contrast theme (percy)', async () => {
+  it.skip('should not have visual regressions in new contrast theme (percy)', async () => {
     await page.goto(url, { waitUntil: ['networkidle2', 'load'] });
     await page.evaluate(() => {
-      document.querySelector('ids-theme-switcher')?.setAttribute('mode', 'contrast');
+      document.querySelector('ids-theme-switcher')?.setAttribute('theme', 'default-contrast');
     });
     await page.waitForTimeout(200);
     await page.evaluate(() => {
       (document.querySelector('#about-example-trigger') as any).click();
     });
     await page.waitForTimeout(200); // approx. time for a Modal to show
-    await percySnapshot(page, 'ids-about-new-contrast');
+    await pageSnapshot(page, 'ids-about-new-contrast');
   });
 });