+++ test/ids-counts/ids-counts-percy-test.ts
@@ -1,31 +1,31 @@
-import percySnapshot from '@percy/puppeteer';
+import pageSnapshot from '../helpers/page-snapshot';
 
 describe('Ids Counts e2e Tests', () => {
   const url = 'http://localhost:4444/ids-counts/example.html';
 
   it('should not have visual regressions in new light theme (percy)', async () => {
     await page.goto(url, { waitUntil: ['networkidle2', 'load'] });
-    await percySnapshot(page, 'ids-counts-new-light');
+    await pageSnapshot(page, 'ids-counts-new-light');
   });
 
   it('should not have visual regressions in standalone css', async () => {
     await page.goto('http://localhost:4444/ids-counts/standalone-css.html', { waitUntil: ['networkidle2', 'load'] });
-    await percySnapshot(page, 'ids-counts-standalone-css', { widths: [1280] });
+    await pageSnapshot(page, 'ids-counts-standalone-css', { widths: [1280] });
   });
 
-  it('should not have visual regressions in new dark theme (percy)', async () => {
+  it.skip('should not have visual regressions in new dark theme (percy)', async () => {
     await page.goto(url, { waitUntil: ['networkidle2', 'load'] });
     await page.evaluate(() => {
-      document.querySelector('ids-theme-switcher')?.setAttribute('mode', 'dark');
+      document.querySelector('ids-theme-switcher')?.setAttribute('theme', 'default-dark');
     });
-    await percySnapshot(page, 'ids-counts-new-dark');
+    await pageSnapshot(page, 'ids-counts-new-dark');
   });
 
-  it('should not have visual regressions in new contrast theme (percy)', async () => {
+  it.skip('should not have visual regressions in new contrast theme (percy)', async () => {
     await page.goto(url, { waitUntil: ['networkidle2', 'load'] });
     await page.evaluate(() => {
-      document.querySelector('ids-theme-switcher')?.setAttribute('mode', 'contrast');
+      document.querySelector('ids-theme-switcher')?.setAttribute('theme', 'default-contrast');
     });
-    await percySnapshot(page, 'ids-counts-new-contrast');
+    await pageSnapshot(page, 'ids-counts-new-contrast');
   });
 });