+++ test/ids-list-view/ids-list-view-percy-test.ts
@@ -1,4 +1,4 @@
-import percySnapshot from '@percy/puppeteer';
+import pageSnapshot from '../helpers/page-snapshot';
 
 describe('Ids List View Percy Tests', () => {
   const url = 'http://localhost:4444/ids-list-view/example.html';
@@ -6,6 +6,6 @@ describe('Ids List View Percy Tests', () => {
   it.skip('should not have visual regressions in new light theme (percy)', async () => {
     await page.goto(url, { waitUntil: ['networkidle2', 'load'] });
     await page.waitForTimeout(200);
-    await percySnapshot(page, 'ids-listview-new-light');
+    await pageSnapshot(page, 'ids-listview-new-light');
   });
 });