+++ test/ids-tabs/ids-tabs-percy-test.ts
@@ -1,4 +1,4 @@
-import percySnapshot from '@percy/puppeteer';
+import pageSnapshot from '../helpers/page-snapshot';
 
 describe('Ids Tabs Percy Tests', () => {
   const url = 'http://localhost:4444/ids-tabs/example.html';
@@ -6,29 +6,29 @@ describe('Ids Tabs Percy Tests', () => {
   it.skip('should not have visual regressions in new light theme (percy)', async () => {
     await page.goto(url, { waitUntil: ['networkidle2', 'load'] });
     await page.waitForSelector('ids-tab[selected]');
-    await percySnapshot(page, 'ids-tabs-new-light');
+    await pageSnapshot(page, 'ids-tabs-new-light');
   });
 
   it('should not have visual regressions in standalone css', async () => {
     await page.goto('http://localhost:4444/ids-tabs/standalone-css.html', { waitUntil: ['networkidle2', 'load'] });
-    await percySnapshot(page, 'ids-tabs-standalone-css', { widths: [960] });
+    await pageSnapshot(page, 'ids-tabs-standalone-css', { widths: [960] });
   });
 
   it.skip('should not have visual regressions in new dark theme (percy)', async () => {
     await page.goto(url, { waitUntil: ['networkidle2', 'load'] });
     await page.waitForSelector('ids-tab[value="contracts"][selected]');
     await page.evaluate(() => {
-      document.querySelector('ids-theme-switcher')?.setAttribute('mode', 'dark');
+      document.querySelector('ids-theme-switcher')?.setAttribute('theme', 'default-dark');
     });
-    await percySnapshot(page, 'ids-tabs-new-dark');
+    await pageSnapshot(page, 'ids-tabs-new-dark');
   });
 
   it.skip('should not have visual regressions in new contrast theme (percy)', async () => {
     await page.goto(url, { waitUntil: ['networkidle2', 'load'] });
     await page.waitForSelector('ids-tab[selected]');
     await page.evaluate(() => {
-      document.querySelector('ids-theme-switcher')?.setAttribute('mode', 'contrast');
+      document.querySelector('ids-theme-switcher')?.setAttribute('theme', 'default-contrast');
     });
-    await percySnapshot(page, 'ids-tabs-new-contrast');
+    await pageSnapshot(page, 'ids-tabs-new-contrast');
   });
 });