+++ test/ids-wizard/ids-wizard-percy-test.ts
@@ -1,27 +1,27 @@
 /* eslint-disable no-underscore-dangle */
-import percySnapshot from '@percy/puppeteer';
+import pageSnapshot from '../helpers/page-snapshot';
 
 describe('Ids Wizard Percy Tests', () => {
   const url = 'http://localhost:4444/ids-wizard/example.html';
 
   it('should not have visual regressions in new light theme (percy)', async () => {
     await page.goto(url, { waitUntil: ['networkidle0', 'domcontentloaded'] });
-    await percySnapshot(page, 'ids-wizard-new-light', { widths: [1280] });
+    await pageSnapshot(page, 'ids-wizard-new-light', { widths: [1280] });
   });
 
   it.skip('should not have visual regressions in new dark theme (percy)', async () => {
     await page.goto(url, { waitUntil: ['networkidle0', 'domcontentloaded'] });
     await page.evaluate(() => {
-      document.querySelector('ids-theme-switcher')?.setAttribute('mode', 'dark');
+      document.querySelector('ids-theme-switcher')?.setAttribute('theme', 'default-dark');
     });
-    await percySnapshot(page, 'ids-wizard-new-dark', { widths: [1280] });
+    await pageSnapshot(page, 'ids-wizard-new-dark', { widths: [1280] });
   });
 
   it.skip('should not have visual regressions in new contrast theme (percy)', async () => {
     await page.goto(url, { waitUntil: ['networkidle0', 'domcontentloaded'] });
     await page.evaluate(() => {
-      document.querySelector('ids-theme-switcher')?.setAttribute('mode', 'contrast');
+      document.querySelector('ids-theme-switcher')?.setAttribute('theme', 'default-contrast');
     });
-    await percySnapshot(page, 'ids-wizard-new-contrast', { widths: [1280] });
+    await pageSnapshot(page, 'ids-wizard-new-contrast', { widths: [1280] });
   });
 });